* Add parity RDATA support, albeit lma0. * Initial window moved back from 1,0 to 0,-1 to sit on transmission group boundary. * Add test for parity RDATA. * Add parity stats. * Transport now keeps buffer for calculating ODATA/RDATA packets. * Transport keeps transmission group shift count to make tg_sqn & pkt_cnt fast to calculate. * Set packet lengths to 16 bit to match wire size. * Transport window transmission group lead packets now store parity counters.

+++ trunk/openpgm/pgm/include/pgm/rxwi.h
@@ -69,7 +69,7 @@ typedef int (*pgm_rxw_commitfn_t)(guint32, gpointer, guint, gpointer);
 
 struct pgm_rxw_packet_t {
 	gpointer        data;
-	guint           length;
+	guint16         length;
 	guint32         sequence_number;
 
 	guint32		apdu_first_sqn;
